Add project for test db fixtures; use it in integration tests (#2907)
* add project for test db fixtures; use it in integration tests
