+++ match/tests/Piipan.Match.Func.ResolutionApi.IntegrationTests/Piipan.Match.Func.ResolutionApi.IntegrationTests.csproj
@@ -24,12 +24,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Content Include="..\..\ddl\match-record.sql" Link="match-record.sql" CopyToOutputDirectory="PreserveNewest" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\..\src\Piipan.Match\Piipan.Match.Func.ResolutionApi\Piipan.Match.Func.ResolutionApi.csproj" />
     <ProjectReference Include="..\..\src\Piipan.Match\Piipan.Match.Core\Piipan.Match.Core.csproj" />
     <ProjectReference Include="..\..\src\Piipan.Match\Piipan.Match.Api\Piipan.Match.Api.csproj" />
+    <ProjectReference Include="..\..\..\shared\src\Piipan.Shared.TestFixtures\Piipan.Shared.TestFixtures.csproj" />
   </ItemGroup>
 </Project>
