+++ shared/shared.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{8195D85F-83A
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Piipan.Shared", "src\Piipan.Shared\Piipan.Shared.csproj", "{2FD5540A-039E-4FCA-AE07-413A6079D0D3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Piipan.Shared.TestFixtures", "src\Piipan.Shared.TestFixtures\Piipan.Shared.TestFixtures.csproj", "{0D459C50-823B-44AA-B1A8-38D99FCB76C9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -48,9 +50,22 @@ Global
 		{2FD5540A-039E-4FCA-AE07-413A6079D0D3}.Release|x64.Build.0 = Release|Any CPU
 		{2FD5540A-039E-4FCA-AE07-413A6079D0D3}.Release|x86.ActiveCfg = Release|Any CPU
 		{2FD5540A-039E-4FCA-AE07-413A6079D0D3}.Release|x86.Build.0 = Release|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Debug|x64.Build.0 = Debug|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Debug|x86.Build.0 = Debug|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Release|x64.ActiveCfg = Release|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Release|x64.Build.0 = Release|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Release|x86.ActiveCfg = Release|Any CPU
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{69F7763A-9006-4165-947A-A35FC5B18C20} = {93135BDE-0152-4513-A6CD-00256F12F847}
 		{2FD5540A-039E-4FCA-AE07-413A6079D0D3} = {8195D85F-83AE-4F7F-B826-F49E316B8C58}
+		{0D459C50-823B-44AA-B1A8-38D99FCB76C9} = {8195D85F-83AE-4F7F-B826-F49E316B8C58}
 	EndGlobalSection
 EndGlobal
