+++ system.sln
@@ -102,9 +102,12 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Piipan.Match", "Piipan.Match", "{207255D2-37B1-410D-91BC-8EBE735088CD}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Piipan.Match.Func.ResolutionApi", "match\src\Piipan.Match\Piipan.Match.Func.ResolutionApi\Piipan.Match.Func.ResolutionApi.csproj", "{FC81EDEF-C422-433E-9F41-E2BCB7A09847}"
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PerformanceTests", "PerformanceTests", "{6A81DE97-0B8C-428C-9D5D-3FC6B3EBF12F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Etl.BulkUpload.Performance.TestRunner", "perf-tests\src\Etl.BulkUpload.Performance.TestRunner\Etl.BulkUpload.Performance.TestRunner.csproj", "{6EF00102-6EEB-4C30-AEF8-73702E7809DD}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{9169E592-EBFA-4340-A7F8-02BBE3ADFBE1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Piipan.Shared.TestFixtures", "shared\src\Piipan.Shared.TestFixtures\Piipan.Shared.TestFixtures.csproj", "{C4C13056-4C2E-4346-AF59-C81FD7D762C2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -252,6 +255,10 @@ Global
 		{6EF00102-6EEB-4C30-AEF8-73702E7809DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6EF00102-6EEB-4C30-AEF8-73702E7809DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6EF00102-6EEB-4C30-AEF8-73702E7809DD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C4C13056-4C2E-4346-AF59-C81FD7D762C2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C4C13056-4C2E-4346-AF59-C81FD7D762C2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C4C13056-4C2E-4346-AF59-C81FD7D762C2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C4C13056-4C2E-4346-AF59-C81FD7D762C2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -301,6 +308,8 @@ Global
 		{207255D2-37B1-410D-91BC-8EBE735088CD} = {28EA778D-49E0-49F6-92E7-56391E11D0E3}
 		{FC81EDEF-C422-433E-9F41-E2BCB7A09847} = {207255D2-37B1-410D-91BC-8EBE735088CD}
 		{6EF00102-6EEB-4C30-AEF8-73702E7809DD} = {6A81DE97-0B8C-428C-9D5D-3FC6B3EBF12F}
+		{9169E592-EBFA-4340-A7F8-02BBE3ADFBE1} = {9D76C087-50BA-499A-BC9C-9029DF0DE7F5}
+		{C4C13056-4C2E-4346-AF59-C81FD7D762C2} = {9169E592-EBFA-4340-A7F8-02BBE3ADFBE1}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7B44BDF3-D32A-4F70-B573-1AD075EEF0EA}
